cambio en los campos de las formas

+++ resources/views/Traficos/TraficosAgregar.blade.php
@@ -230,7 +230,7 @@
                                 message: 'El costo debe tener como máximo 255 caracteres'
                             },
 
-                            integer:{
+                            numeric:{
                                 message: 'Ingresa un número'
                             }
                         }
